New tool added: Universal 5.3.0.xxxx patcher
You can patch your client to play Arctium, no matter which version or
architecture it is,whenever it is 5.3.0.xxxx and for Windows.
Just click Tools->Client Patcher and select your .exe file (either
Wow.exe or Wow-64.exe).
It will detect if x64 or x32 and apply the right patches, no matter
which build your client is (tested up to 16992).